% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db.R
\name{add_to_database}
\alias{add_to_database}
\title{Add selected data to a MySql database}
\usage{
add_to_database(db, leagueID = NULL, stats = FALSE, league = FALSE,
  weeks = c(), all = FALSE, progressbar = FALSE)
}
\arguments{
\item{db}{open database object returned from dbConnect}

\item{leagueID}{Numeric league identifier}

\item{stats}{Boolean.  If true update stats table}

\item{league}{Boolean.  If true update league table}

\item{weeks}{vector of weeks to collect data for}

\item{all}{Boolean.  If true collect data for all weeks (overrides weeks vector)}

\item{progressbar}{Boolean.  If true a progress bar is displayed}
}
\description{
Retrieve fantasy football data from fantasy.premierleague.com site and
add to a MySQL database
}

