% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{captainChoice}
\alias{captainChoice}
\title{Was the best captain selected?}
\usage{
captainChoice(l, weeks = c(), managers = c())
}
\arguments{
\item{l}{List of obtained from read_database}

\item{weeks}{Vector of weeks.  If empty then include all weeks}

\item{managers}{Vector of teams.  Vector of manager names, manager IDs or team
names.  If empty then include all teams}
}
\value{
dataframe containing table
}
\description{
Retrieve a table indicating points differential between the captain chosen and the
    best choice
}

