% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPL.R
\name{.teamID}
\alias{.teamID}
\title{.teamID}
\usage{
.teamID(fpl, id)
}
\arguments{
\item{fpl}{Output from getFPLData}

\item{Team}{identifier (single number or vector of numbers)}
}
\description{
Return short Premier League team name from ID
}
