% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{.week_dropped}
\alias{.week_dropped}
\title{week_dropped}
\usage{
.week_dropped(player = NA, team = NA, week_added = NA,
  max_weeks = NA, data = NA)
}
\arguments{
\item{player.}{Player's name}

\item{team.}{Player's team}

\item{week_added.}{Initial week to use for search - player dropped at some point after this week}

\item{max_weeks.}{Week to return if player has not been dropped (typically maximum week of data)}
}
\value{
.  Integer.  Week player dropped or `max_weeks` if player has not been dropped
}
\description{
helper function to determine the week a player was dropped
}
