% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{extract_table}
\alias{extract_table}
\title{Return requested table from the bootstrap data}
\usage{
extract_table(bootstrap = get_bootstrap(), table = "elements")
}
\arguments{
\item{bootstrap}{Data from FPL API}

\item{table}{Name of table to return}
}
\value{
a tibble
}
\description{
Return requested table from the bootstrap data
}
