% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{fixtures}
\alias{fixtures}
\title{Summary data for fixtures in the current FPL season}
\usage{
fixtures()
}
\description{
Gives a tibble containing data on all fixtures in the current FPL season.
}
\examples{
fixtures()
}
