% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\name{fpl_get_bootstrap}
\alias{fpl_get_bootstrap}
\title{Get data from bootstrap}
\usage{
fpl_get_bootstrap(
  url = "https://fantasy.premierleague.com/api/bootstrap-static/"
)
}
\arguments{
\item{url}{Address of FPL API}
}
\value{
a tibble
}
\description{
Get data from bootstrap
}
