% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{fpl_get_fixtures}
\alias{fpl_get_fixtures}
\title{Retrieve fixture data for the current FPL season}
\usage{
fpl_get_fixtures()
}
\value{
a tibble
}
\description{
Retrieve fixture data for the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/drf/bootstrap-static}{bootstrap-static JSON}.
}
\examples{
fixtures <- fpl_get_fixtures()
}
