% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{fpl_get_fixtures_gameweek}
\alias{fpl_get_fixtures_gameweek}
\title{Get detailed data on a gameweek's fixtures}
\usage{
fpl_get_fixtures_gameweek(gameweek)
}
\value{
a tibble
}
\description{
Retrieve detailed data on a gameweek's fixtures for the current FPL season,
obtained via the
\href{https://fantasy.premierleague.com/drf/fixtures/?event=1}{fixtures event endpoint}.
}
\examples{
\donttest{
fpl_get_fixtures_gameweek(gameweek = 8)
}
}
