% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweeks.R
\name{fpl_get_gameweek_current}
\alias{fpl_get_gameweek_current}
\title{Get current gameweek}
\usage{
fpl_get_gameweek_current()
}
\value{
numeric vector
}
\description{
Retrieve current gameweek via the
 \href{https://fantasy.premierleague.com/api/bootstrap-static}{bootstrap-static endpoint}.
}
\examples{
\donttest{
fpl_get_gameweek_current()
}
}
