% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweeks.R
\name{fpl_get_gameweek_next}
\alias{fpl_get_gameweek_next}
\title{Get next gameweek}
\usage{
fpl_get_gameweek_next()
}
\value{
numeric vector
}
\description{
Retrieve next gameweek via the
 \href{https://fantasy.premierleague.com/api/bootstrap-static}{bootstrap-static endpoint}.
}
\examples{
\donttest{
fpl_get_gameweek_next()
}
}
