% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweeks.R
\name{fpl_get_gameweeks}
\alias{fpl_get_gameweeks}
\title{Get data on gameweeks}
\usage{
fpl_get_gameweeks()
}
\value{
a tibble
}
\description{
Retrieve gameweeks data for the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/drf/events/}{events endpoint}.
}
\examples{
\donttest{
fpl_get_gameweeks()
}
}
