% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{fpl_get_leagues_classic}
\alias{fpl_get_leagues_classic}
\title{Get data on a classic league}
\usage{
fpl_get_leagues_classic(league_id)
}
\value{
a list
}
\description{
Retrieve a classic league's data for the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/drf/leagues-classic-standings/1137}{leagues classic standings endpoint}.
}
\examples{
\donttest{
fpl_get_leagues_classic(league_id = 1137)
}
}
