% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{fpl_get_phases}
\alias{fpl_get_phases}
\title{Get data on phases}
\usage{
fpl_get_phases()
}
\value{
a tibble
}
\description{
Retrieve phases data for the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/drf/phases/}{phases endpoint}.
}
\examples{
\donttest{
fpl_get_phases()
}
}
