% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{fpl_get_player_current}
\alias{fpl_get_player_current}
\title{Get current season data on a player}
\usage{
fpl_get_player_current(player_id, convert_prices = TRUE)
}
\arguments{
\item{player_id}{player's ID (\code{id} field from \code{\link{fpl_get_player_all}} output).}

\item{convert_prices}{Convert player price-related metrics to familiar denominations
(TRUE by default)}
}
\value{
a tibble
}
\description{
Retrieve gameweek-level data for a player in the current FPL season,
obtained via the
\href{https://fantasy.premierleague.com/api/element-summary/1}{element summary endpoint}.
}
\examples{
\donttest{
fpl_get_player_current(player_id = 3)
}
}
