% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{fpl_get_player_detailed}
\alias{fpl_get_player_detailed}
\title{Retrieve detailed data for a player in the current FPL season}
\usage{
fpl_get_player_detailed(player_id)
}
\arguments{
\item{player_id}{\code{id} field from \code{\link{fpl_get_players}}.}
}
\value{
a tibble
}
\description{
Retrieve detailed data for a player in the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/drf/bootstrap-static}{bootstrap-static JSON}.
}
\examples{
\dontrun{
fpl_get_player_detailed(player_id = 2)
}
}
