% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{fpl_get_player_seasons}
\alias{fpl_get_player_seasons}
\title{Retrieve historic seasons summary data for a player in the current FPL season}
\usage{
fpl_get_player_seasons(player_id)
}
\arguments{
\item{player_id}{\code{id} field from \code{\link{fpl_get_players}}.}
}
\value{
a tibble
}
\description{
Retrieve summary (season-level) data for a player in the current FPL season, for all
previous FPL seasons, obtained via the
\href{https://fantasy.premierleague.com/drf/bootstrap-static}{bootstrap-static JSON}.
N.B. the player must have competed in a previous FPL season for this function to return
data.
}
\examples{
\dontrun{
fpl_get_player_seasons(player_id = 3)
}
}
