% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{fpl_get_player_types}
\alias{fpl_get_player_types}
\title{Get metadata on player types}
\usage{
fpl_get_player_types()
}
\value{
a tibble
}
\description{
Retrieve player type metadata for the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/api/element-types/}{element types endpoint}.
}
\examples{
\donttest{
fpl_get_player_types()
}
}
