% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{fpl_get_players}
\alias{fpl_get_players}
\title{Retrieve player data for the current FPL season}
\usage{
fpl_get_players()
}
\value{
a tibble
}
\description{
Retrieve player data for the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/drf/bootstrap-static}{bootstrap-static JSON}.
}
\examples{
players <- fpl_get_players()
}
