% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{fpl_get_user_gameweek}
\alias{fpl_get_user_gameweek}
\title{Retrieve an FPL user's data for a gameweek in the current FPL season}
\usage{
fpl_get_user_gameweek(user_id, gameweek)
}
\arguments{
\item{user_id}{FPL user ID (retrieve from a user's team page on the \href{https://fantasy.premierleague.com}{FPL site}).}

\item{gameweek}{FPL gameweek (1-38).}
}
\value{
a list
}
\description{
Retrieve an FPL user's gameweek player picks, obtained via the
\href{https://fantasy.premierleague.com/drf/bootstrap-static}{bootstrap-static JSON}.
N.B. the current season must have started for this function to return data.
}
\examples{
\dontrun{
fpl_get_user_gameweek(user_id = 123, gameweek = 1)
}

}
