% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{fpl_get_user_leagues}
\alias{fpl_get_user_leagues}
\title{Get league data on a user}
\usage{
fpl_get_user_leagues(user_id)
}
\arguments{
\item{user_id}{user ID (found in a user's team page URL on the \href{https://fantasy.premierleague.com}{main FPL site}).}
}
\value{
a list
}
\description{
Retrieve league information for a user in the current FPL season,
obtained via the
\href{https://fantasy.premierleague.com/drf/entry/3808385}{user entry history endpoint}.
}
\examples{
fpl_get_user_leagues(user_id = 3808385)
}
