% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{fpl_get_user_performance}
\alias{fpl_get_user_performance}
\title{Retrieve an FPL user's performance data}
\usage{
fpl_get_user_performance(user_id)
}
\arguments{
\item{user_id}{FPL user ID (retrieve from a user's team page on the \href{https://fantasy.premierleague.com}{FPL site}).}
}
\value{
a list
}
\description{
Retrieve an FPL user's performance data, obtained via the
\href{https://fantasy.premierleague.com/drf/bootstrap-static}{bootstrap-static JSON}.
}
\examples{
fpl_get_user_performance(user_id = 123)
}
