% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{fpl_league}
\alias{fpl_league}
\title{Get league table}
\usage{
fpl_league(f, max_week = 0, out_type = "total")
}
\arguments{
\item{f}{an fpl object}

\item{max_week}{Maximum week.  Defaults to maximum in database}

\item{out_type}{Type of output.  If total then return the total to the maximum week.
If cumulative then return a cumulative score by week}
}
\value{
dataframe containing table
}
\description{
Retrieve a cumulative league table a fantasy football league
}

