% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{fpl_league_weekly}
\alias{fpl_league_weekly}
\title{Get weekly scores}
\usage{
fpl_league_weekly(l, weeks = c())
}
\arguments{
\item{l}{List of obtained from read_database}

\item{weeks}{Vector of weeks.  If empty then include all weeks}
}
\value{
dataframe containing table
}
\description{
Retrieve weekly scores for a fantasy football league
}

