% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplr-package.R
\docType{package}
\name{fplr-package}
\alias{fplr}
\alias{fplr-package}
\title{fplr: Tools for Working with Fantasy Premier League (FPL) Data in R}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Tools for Fantasy Premier League (FPL, see <https://fantasy.premierleague.com> for info) 
    users to retrieve, and work with, FPL data using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ewenme.github.io/fplr}
  \item \url{https://github.com/ewenme/fplr}
  \item Report bugs at \url{https://github.com/ewenme/fplr/issues}
}

}
\author{
\strong{Maintainer}: Ewen Henderson \email{ewenhenderson@gmail.com} (0000-0002-4748-4693)

}
\keyword{internal}
