% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_bootstrap}
\alias{get_bootstrap}
\title{Get data from bootstrap}
\usage{
get_bootstrap(url = "https://fantasy.premierleague.com/api/bootstrap-static/")
}
\arguments{
\item{url}{Address of FPL API}
}
\value{
a list of data frames
}
\description{
Get data from bootstrap
}
