% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gw_data.R
\name{get_gw_data}
\alias{get_gw_data}
\title{Function to extract data for a specific player(s) from a certain gameweek or weeks.}
\usage{
get_gw_data(players, gw = 1:38)
}
\arguments{
\item{players}{a vector of players, referred to either using their ID's (numerical
vector) or names (character vector).
as they are written in the id_lookup data frame. I would reccomend using the players
ID's due to potential issues with special characters.}

\item{gw}{The FPL Gameweek to pull data for, a vector of integers between 1 and 38.
Note that if gameweeks chosen haven't been played yet then data won't be displayed for them.
Defaults to all gameweeks.}
}
\value{
A data frame, with one row of data for each chosen player for each chosen gameweek.
The data has 29 variables:
\describe{
   \item{name}{Player name}
   \item{opponent}{Team played against in this round}
   \item{total_points}{Total fpl points earned in this gameweek}
   \item{was_home}{Boolean - did the player play at home?}
   \item{team_h_score, team_a_score}{Number of goals scored by the home or away team}
   \item{round}{The Gameweek}
   \item{minutes}{Minutes played by the player in this gameweek}
   \item{goals_scored, assists, clean_sheets, goals_conceded, own_goals, penalties_saved,
   penalties_missed, yellow_cards, red_cards, saves}{Statistics related to the players performance in this gameweek}
   \item{bonus}{fpl bonus points awarded to the player in this gameweek}
   \item{bps}{Points allocated to the player by the fpl bonus points system - used to determine how many bonus point player should be awarded}
   \item{influence, creativity,threat, ict_index}{Values calculated by fpl algorithms}
   \item{value}{The cost of the player in this gameweek  (£m)}
   \item{transfers_balance, transfers_in, transfers_out}{Figures relating to how many managers transferred this player in/out for this gameweek}
   \item{selected}{Number of fpl managers selected by}}
}
\description{
Function to extract data for a specific player(s) from a certain gameweek or weeks.
}
\examples{
 # Not run since it's hard to pick a player who'll be around
 # forever
 #get_gw_data('Simon Mignolet')

get_gw_data(190, 1:2)

get_gw_data(1:10, 3)

#get_gw_data(c('Simon Mignolet', 'Laurent Koscielny'))
}
