% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_data.R
\name{get_player_data}
\alias{get_player_data}
\title{Function to extract data for a specific player from a certain gameweek.}
\usage{
get_player_data(players, gw = 1:38)
}
\arguments{
\item{players}{a vector of players, referred to either using their ID's (numerical
vector) or names (character vector).
as they are written in the id_lookup data frame. I would reccomend using the players
ID's due to potential issues with special characters.}

\item{gw}{The FPL Gameweek to pull data for, a vector of integers between 1 and 38.
Note that if gameweeks chosen haven't been played yet then data won't be displayed for them.
Defaults to all gameweeks.}
}
\value{
A data frame, with one row of data for each chosen player for each chosen gameweek
}
\description{
Function to extract data for a specific player from a certain gameweek.
}
\examples{
get_player_data('Simon Mignolet')

get_player_data(190, 1:2)

get_player_data(1:10, 3)

get_player_data(c('Simon Mignolet', 'Laurent Koscielny'))
}
