% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_data.R
\name{get_season_data}
\alias{get_season_data}
\title{Function which gives access to higher level data about each player, aggregated at a
season level.}
\usage{
get_season_data(players)
}
\arguments{
\item{players}{Either a character vector of player names or a numeric vector of player
ID's which can be obtained from the `id_lookup` table}
}
\value{
A data frame with one row for each player chosen and 27 variables as follows:
\describe{
   \item{player_name, team_name}{Information about the player}
   \item{start_cost}{Cost (in £m) at the start of the season (or when introduced to the game)}
   \item{now_cost}{Cost at the time of calling the function (in £m)}
   \item{dreamteam_count}{How many times this player has appeared in the dreamteam (team of the gameweek)}
   \item{form}{As calculated by pls based on games in the last 30 days}
   \item{points_per_game}{Average points per Gameweek participated in}
   \item{selected_by_percent}{Percent of managers who's team this player is selected in (at the time of running the function)}
   \item{status}{One of: Available, Injured - won't play, Injured - may play, Suspended or Unavailable}
   \item{total_points}{Total fpl points acrued by this player over the season to date}
   \item{goals_scored, assists, clean_sheets, goals_conceded, own_goals, penalties_saved,
   penalties_missed, yellow_cards, red_cards, saves}{Statistics relating to the players performance over the whole season}
   \item{bonus}{Total fpl bonus points earned over the whole season}
   \item{bps}{Points awarded to this player by the fpl bonus point system over the season}
   \item{influence, creativity, threat, ict_index}{Current values from fpl algorithms}}
}
\description{
Function which gives access to higher level data about each player, aggregated at a
season level.
}
\examples{
# Don't run as will fail for missing players in the future
#get_season_data(301)

#get_season_data('John Fleck')

#get_season_data(350:352)
}
