% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPL.R
\name{getEntry}
\alias{getEntry}
\title{Get an entry info}
\usage{
getEntry(entry = NULL)
}
\arguments{
\item{entry}{Numeric entry identifier}
}
\description{
Return entry details and historical league information
}

