% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPL.R
\name{getFPLData}
\alias{getFPLData}
\title{Get team, player and game data}
\usage{
getFPLData()
}
\value{
a list containing phases, elements, game-settings, total-players, teams,
element_types and events
}
\description{
Retrieve fantasy football data from fantasy.premierleague.com site.
}
