% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPL.R
\name{getLeagueEntries}
\alias{getLeagueEntries}
\title{Get all entries for a particular league}
\usage{
getLeagueEntries(leagueID, wk = 1, fplData = NULL)
}
\arguments{
\item{leagueID}{Numeric league identifier}

\item{wk}{Week number}

\item{fplData}{Output from getFPLData}
}
\description{
Return a list of dataframes each containing an entry
}

