% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPL.R
\name{getLeagueTable}
\alias{getLeagueTable}
\title{Get a league table}
\usage{
getLeagueTable(leagueID = NULL, wk = 1)
}
\arguments{
\item{leagueID}{Numeric league identifier}

\item{wk}{Week number}
}
\description{
Retrieve a fantasy football league table for a specific week
}

