% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerDetailed.R
\name{getPlayerDetailed}
\alias{getPlayerDetailed}
\title{getPlayerDetailed}
\usage{
getPlayerDetailed(player_id)
}
\arguments{
\item{player_id}{Player ID. Also see \code{getPlayerId} function.}
}
\description{
Returns a tibble containing round-by-round data for a given player in the current FPL season.
}
\examples{
getPlayerDetailed(1)
getPlayerDetailed(54)
}
\keyword{player}
