% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerHistory.R
\name{getPlayerHistory}
\alias{getPlayerHistory}
\title{getPlayerHistory}
\usage{
getPlayerHistory(player_id)
}
\arguments{
\item{player_id}{Player ID.}
}
\description{
Returns a tibble containing historical seasons data for a player in the current FPL season.
}
\examples{
getPlayersHistory()
}
\keyword{players}
