% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerId.R
\name{getPlayerId}
\alias{getPlayerId}
\title{getPlayerId}
\usage{
getPlayerId(player_name, team_name)
}
\arguments{
\item{player_name}{Name of player.}

\item{team_name}{Name of team.}
}
\description{
Returns the player_id for a given FPL player in the current FPL season. This is used in calls to other functions (see \code{getPlayerDetailed} and \code{getPlayerSeasons}).
}
\examples{
getPlayerId("Petr Cech", "Arsenal")
}
\keyword{player}
