% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayers.R
\name{getPlayers}
\alias{getPlayers}
\title{getPlayers}
\usage{
getPlayers()
}
\description{
Returns a tibble containing the FPL player data for the current FPL season.
}
\examples{
getPlayers()
}
\keyword{players}
