% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerSeasons.R
\name{getPlayerSeasons}
\alias{getPlayerSeasons}
\title{getPlayerSeasons}
\usage{
getPlayerSeasons(player_id)
}
\arguments{
\item{player_id}{Player ID. Also see \code{getPlayerId} function.}
}
\description{
Returns a tibble containing a historical season overview for a given player in the current FPL season.
}
\examples{
getPlayerSeasons(1)
getPlayerSeasons(54)
}
\keyword{player}
