% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPL.R
\name{getTeam}
\alias{getTeam}
\title{Get team data}
\usage{
getTeam(entry = NULL, wk = 1)
}
\arguments{
\item{entry}{Numeric entry identifier}

\item{wk}{Week number}
}
\description{
Return team information from a specific week.
}

