% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id_lookup}
\alias{id_lookup}
\title{Corresponding id, team and position for every player in fpl 2019/20}
\format{A data frame with 532 rows and 4 variables:
\describe{
  \item{id}{unique id for each player, used to extract player specific data}
  \item{name}{the name of the player}
  \item{position}{position player is classified as on fpl}
  \item{team_name}{name of the team the player plays for}
}}
\source{
\url{https://fantasy.premierleague.com/api/bootstrap-static/}
}
\usage{
id_lookup
}
\description{
A dataset containing the basic information of over 500 players.
}
\keyword{datasets}
