% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db.R
\name{initialize_database}
\alias{initialize_database}
\title{Add initial data to a MySql database}
\usage{
initialize_database(db, leagueID = NULL)
}
\arguments{
\item{db}{open database object returned from dbConnect}

\item{leagueID}{Numeric league identifier}
}
\description{
Retrieve fantasy football data from fantasy.premierleague.com site and
add to a MySQL database
}

