% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{keeperChoice}
\alias{keeperChoice}
\title{Was the best keeper selected?}
\usage{
keeperChoice(f, weeks = c(), managers = c())
}
\arguments{
\item{f}{an fpl object}

\item{weeks}{Vector of weeks.  If empty then include all weeks}

\item{managers}{Vector of teams.  Vector of manager names, manager IDs or team
names.  If empty then include all teams}
}
\value{
dataframe
}
\description{
Retrieve a table indicating points differential between the keeper chosen and the
    best choice
}
