% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPL.R
\name{playerCount}
\alias{playerCount}
\title{Identify players included in and missing from fantasy teams}
\usage{
playerCount(l.teams, fpl)
}
\arguments{
\item{l.teams}{A series of teams as generated by getLeagueEntries}

\item{fpl}{Output from getFPLData}
}
\description{
Identify players included in and missing from fantasy teams
}

