% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerDetailed.R
\name{playerDetailed}
\alias{playerDetailed}
\title{Detailed data for a player in the current FPL season}
\usage{
playerDetailed(player_id)
}
\arguments{
\item{player_id}{\code{id} field from \code{\link{players}} tibble for a desired player.}
}
\description{
Returns a tibble containing gameweek-level data for a given player in the current FPL season.
}
\examples{
playerDetailed(player_id = 1)
playerDetailed(player_id = 54)
}
