% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerNames.R
\name{playerNames}
\alias{playerNames}
\title{playerNames}
\usage{
playerNames()
}
\description{
Returns the current FPL season's player names, along with their team and squad number. This is used in calls to other functions (see \code{getPlayerId}).
}
\examples{
playerNames()
}
\keyword{players}
