% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{players}
\alias{players}
\title{Summary player data for the current FPL season}
\usage{
players()
}
\description{
Gives a tibble containing summary data for all players in the current FPL season.
}
\examples{
players()
}
