% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players1516.R
\docType{data}
\name{players1516}
\alias{players1516}
\title{players1516}
\format{An object of class \code{tibble}.}
\usage{
data(players1516)
}
\description{
Summary data for all players in the 2015/16 FPL season.
}
\examples{
data(players1516)
}
\keyword{datasets}
