% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players1617.R
\docType{data}
\name{players1617}
\alias{players1617}
\title{players1617}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 683 rows and 34 columns.}
\source{
\url{https://fantasy.premierleague.com/drf/bootstrap-static}, downloaded 2017-06-06
}
\usage{
players1617
}
\description{
Summary data for all players in the 2016/17 FPL season.
}
\examples{
data(players1617)
}
\keyword{datasets}
