% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerSeasons.R
\name{playerSeasons}
\alias{playerSeasons}
\title{Historic FPL seasons overview for a player in the current season}
\usage{
playerSeasons(player_id)
}
\arguments{
\item{player_id}{\code{id} field from \code{\link{players}} tibble for a desired player.}
}
\description{
Returns a tibble containing a season history overview for a given player in the current FPL season.
}
\examples{
playerSeasons(player_id = 1)
}
