% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{points_by_week}
\alias{points_by_week}
\title{Internal function to get weekly scores}
\usage{
points_by_week(f, weeks = c(), inc_transfers)
}
\arguments{
\item{f}{an fpl object}

\item{weeks}{Vector of weeks.  If empty then include all weeks}

\item{inc_transfsers}{Boolean.  Should points lost by transfers be taken into account.  A value
of TRUE subtracts points from weekly totals when transfers cost points}
}
\value{
dataframe containing table
}
\description{
Retrieve weekly scores for a fantasy football league
Used by fpl_league_weekly and fpl_league functions
}

