% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{points_on_bench}
\alias{points_on_bench}
\title{Determine points left on bench}
\usage{
points_on_bench(f, weeks = c(), managers = c())
}
\arguments{
\item{f}{an fpl object}

\item{weeks}{Vector of weeks.  If empty then include all weeks}

\item{managers}{Vector of teams.  Vector of manager names, manager IDs or team
names.  If empty then include all teams}
}
\value{
list of two dataframes
}
\description{
Calculate number of points left on bench as total and as percentage of weekly score
}
