% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPL.R
\name{pointsFrame}
\alias{pointsFrame}
\title{Player Point data frame}
\usage{
pointsFrame(fpl = NULL, maxWeek = NULL)
}
\arguments{
\item{fpl}{Output from getFPLData}

\item{maxWeek}{maximum week number to run (current week if set to NULL)}
}
\description{
Return a data frame of players and points up to a maximum week
}

