% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db.R
\name{read_database}
\alias{read_database}
\title{Read data back from a MySQL database}
\usage{
read_database(db)
}
\arguments{
\item{db}{open database object returned from dbConnect}
}
\value{
named list of database tables
}
\description{
Retrieve all tables fpr subsequent analysis
}

