% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{substitution_analysis}
\alias{substitution_analysis}
\title{Substitution Analysis}
\usage{
substitution_analysis(f, start_week = 2, number_weeks = 1, managers = c())
}
\arguments{
\item{f}{an fpl object}

\item{start_week}{Integer.  Week to start the analysis from (default = 2)}

\item{number_weeks}{Integer.  Number of weeks to run analysus (default = current week).}

\item{managers}{Vector of teams.  Vector of manager names, manager IDs or team
names.  If empty then include all teams}
}
\description{
Analyze substitutions - how beneficial were the substitutions?
Take the substitutions applied in a game week and project out a number of weeks to determine
the points differential between the added player and the dropped player.
}

