% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{substitution_analysis_all}
\alias{substitution_analysis_all}
\title{Substitution Analysis}
\usage{
substitution_analysis_all(f, managers = c())
}
\arguments{
\item{f}{an fpl object}

\item{managers}{Vector of teams.  Vector of manager names, manager IDs or team
names.  If empty then include all teams}
}
\description{
Analyze substitutions - how beneficial were the substitutions?
Take all substitutions and project out until the substituted player was dropped to determine
the points differential between the added player and the dropped player.
}
