% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{substitutions}
\alias{substitutions}
\title{Substitutions by week}
\usage{
substitutions(f, weeks = c(), managers = c())
}
\arguments{
\item{f}{an fpl object}

\item{weeks}{Vector of weeks.  If empty then include all weeks.
If a vector of length one then determine substitutes up to this week.
If a vector of length two then determine substitutes between these weeks.}

\item{managers}{Vector of teams.  Vector of manager names, manager IDs or team
names.  If empty then include all teams}
}
\value{
list of three dataframes.  The first contains all data and the second contains a summary table
    #' \itemize{
      \item substitute data - team, week, position, player out (and team id), player in (and team id)
      \item summary of subs by weekly count
      \item summary of subs by player names (out -> in)
    }
}
\description{
Return a table of substitutions by manager by week
}

