% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{teamIDs}
\alias{teamIDs}
\title{Return a list of teams}
\usage{
teamIDs(l, teams = c())
}
\arguments{
\item{l}{List of obtained from read_database}

\item{teams}{Vector of teams.  Vector of manager names, manager IDs or team
names.  If empty then include all teams}
}
\value{
List of team IDs
}
\description{
Return a list of team IDs from a list of manager names, team names or simply the IDs
    themselves.
}

