% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teams}
\alias{teams}
\title{Data for each PL team}
\format{A data frame with 20 rows and 10 variables:
\describe{
   \item{id}{id for each team}
   \item{name}{full name}
   \item{short_name}{abbreviated name}
   \item{strength}{Overall strenght value assigned by FPL algorithm}
   \item{stength_*}{Breakdown of a teams strengths, split in to overall, attacking
   and defending for home and away}
}}
\source{
\url{https://fantasy.premierleague.com/api/bootstrap-static/}
}
\usage{
teams
}
\description{
Mainly used to convert ID's to actual team names, but the strength assigned
by the FPL has been left in just in case it is of any interest.
}
\keyword{datasets}
