% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{use_chip}
\alias{use_chip}
\title{List chip usage}
\usage{
use_chip(f, managers = c())
}
\arguments{
\item{f}{an fpl object}

\item{managers}{Vector of teams.  Vector of manager names, manager IDs or team
names.  If empty then include all teams}
}
\value{
dataframe containing weekly table of chips
}
\description{
Return a data frame of weeks when chips have been used
}

