% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{userHistory}
\alias{userHistory}
\title{Get previous season's performance for an FPL user}
\usage{
userHistory(user_id)
}
\arguments{
\item{user_id}{FPL user ID (retrieve from the corresponding user's \url{https://fantasy.premierleague.com} team page URL).}
}
\description{
Gives a tibble containing previous FPL season's performance info for a given user.
}
\examples{
userHistory(user_id = 123)
}
