% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_db_fns.R
\name{week_score}
\alias{week_score}
\title{Internal function to return a weekly total for a grouped team and week}
\usage{
week_score(d, inc_transfers, entry_use)
}
\arguments{
\item{d}{Data frame for a specific manager and week}

\item{inc_transfers}{Boolean.  Should points lost by transferring players be taken into account?}

\item{entry_use}{Dataframe.  A data frame containing the entry_weeks table as obtained from
read_database.  It requires that the two columns be remaned (use_entry = entry, use_week = week)}
}
\description{
Internal function to return a weekly total for a grouped team and week
}
