% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{fpl_get_bootstrap}
\alias{fpl_get_bootstrap}
\title{Retrieve FPL bootstrap-static JSON in full}
\usage{
fpl_get_bootstrap()
}
\value{
a list
}
\description{
Retrieve FPL \href{https://fantasy.premierleague.com/drf/bootstrap-static}{bootstrap-static JSON}
JSON in full.
}
\examples{
fpl_get_bootstrap()
}
