% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{fpl_get_bootstrap}
\alias{fpl_get_bootstrap}
\title{Get bootstrap-static data}
\usage{
fpl_get_bootstrap()
}
\value{
a list
}
\description{
Retrieve all available data via the
 \href{https://fantasy.premierleague.com/api/bootstrap-static}{bootstrap-static API endpoint}.
}
\examples{
\donttest{
fpl_get_bootstrap()
}
}
