% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{fpl_get_bootstrap}
\alias{fpl_get_bootstrap}
\title{Get FPL's bootstrap-static data}
\usage{
fpl_get_bootstrap()
}
\value{
a list
}
\description{
Retrieve FPL's \href{https://fantasy.premierleague.com/drf/bootstrap-static}{bootstrap-static}
JSON data in full.
}
\examples{
\donttest{
fpl_get_bootstrap()
}
}
