% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{fpl_get_fixtures}
\alias{fpl_get_fixtures}
\title{Get data on fixtures}
\usage{
fpl_get_fixtures()
}
\value{
a tibble
}
\description{
Retrieve data on fixtures for the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/drf/fixtures/}{fixtures endpoint}.
}
\examples{
fpl_get_fixtures()
}
