% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{fpl_get_players}
\alias{fpl_get_players}
\title{Get data on all players}
\usage{
fpl_get_players()
}
\value{
a tibble
}
\description{
Retrieve player data for the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/drf/elements/}{elements endpoint}.
}
\examples{
\donttest{
fpl_get_player_all()
}
}
