% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{fpl_get_stats_lookup}
\alias{fpl_get_stats_lookup}
\title{Get stats variable name/key values}
\usage{
fpl_get_stats_lookup()
}
\value{
a tibble
}
\description{
Retrieve statistic variable name and key values via the
 \href{https://fantasy.premierleague.com/api/bootstrap-static}{bootstrap-static API endpoint}.
}
\examples{
\donttest{
fpl_get_stats_lookup()
}
}
