% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{fpl_get_teams}
\alias{fpl_get_teams}
\title{Get data on teams}
\usage{
fpl_get_teams()
}
\value{
a tibble
}
\description{
Retrieve data on teams in the current FPL season, obtained via the
\href{https://fantasy.premierleague.com/drf/teams}{teams endpoint}.
}
\examples{
fpl_get_teams()
}
