% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPL.R
\name{getLeagueTable}
\alias{getLeagueTable}
\title{Get a league table}
\usage{
getLeagueTable(leagueID = NULL, phase = NULL)
}
\arguments{
\item{leagueID}{Numeric league identifier}

\item{phase}{Phase number.  If \code{NULL} then return the current league table}
}
\description{
Retrieve a fantasy football league table for a specific phase
}

