% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerDetailed.R
\name{getPlayerDetailed}
\alias{getPlayerDetailed}
\title{getPlayerDetailed}
\usage{
getPlayerDetailed(player_id)
}
\arguments{
\item{player_id}{Player ID.}
}
\description{
Returns a tibble containing round-by-round data for a player in the current FPL season.
}
\examples{
getPlayerDetailed()
}
\keyword{player}
