% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerDetailed.R
\name{getPlayerDetailed}
\alias{getPlayerDetailed}
\title{getPlayerDetailed}
\usage{
getPlayerDetailed(player_id)
}
\arguments{
\item{player_id}{\code{id} field from \code{getPlayers()} tibble for a desired player.}
}
\description{
Returns a tibble containing round-by-round data for a given player in the current FPL season.
}
\examples{
getPlayerDetailed(player_id = 1)
getPlayerDetailed(player_id = 54)
}
\keyword{player}
