% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerId.R
\name{getPlayerId}
\alias{getPlayerId}
\title{getPlayerId}
\usage{
getPlayerId(player_name, team_name)
}
\arguments{
\item{player_name}{Name of player.}

\item{team_name}{Name of team.}
}
\description{
Returns the player ID for an FPL player in the current FPL season.
}
\examples{
getPlayerId()
}
\keyword{player}
