% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayers.R
\name{getPlayers}
\alias{getPlayers}
\title{getPlayers}
\usage{
getPlayers()
}
\description{
Gives a tibble containing summary data for all players in the current FPL season.
}
\examples{
getPlayers()
}
\keyword{players}
