% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerSeasons.R
\name{getPlayerSeasons}
\alias{getPlayerSeasons}
\title{Historic FPL seasons overview for a player in the current season}
\usage{
getPlayerSeasons(player_id)
}
\arguments{
\item{player_id}{\code{id} field from \code{\link{getPlayers}} tibble for a desired player.}
}
\description{
Returns a tibble containing a season history overview for a given player in the current FPL season.
}
\examples{
getPlayerSeasons(player_id = 1)
getPlayerSeasons(player_id = 54)
}
