% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{teams}
\alias{teams}
\title{Summary data for teams in the current FPL season}
\usage{
teams()
}
\description{
Gives a tibble containing summary data for all teams in the current FPL season.
}
\examples{
teams()
}
