% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{userInfo}
\alias{userInfo}
\title{Get information on an FPL user}
\usage{
userInfo(user_id)
}
\arguments{
\item{user_id}{FPL user ID (retrieve from the corresponding user's \url{https://fantasy.premierleague.com} team page URL).}
}
\description{
Gives a tibble containing summary info for a given user in the current FPL season.
}
\examples{
userInfo(user_id = 123)
}
