% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{userPerformance}
\alias{userPerformance}
\title{Detailed user performance data for the current FPL season.}
\usage{
userPerformance(user_id)
}
\arguments{
\item{user_id}{FPL user ID (retrieve from the corresponding user's \url{https://fantasy.premierleague.com} team page URL).}
}
\description{
Returns a tibble containing round-by-round performance info for a given user in the current FPL season.
}
\examples{
userPerformance(user_id = 123)
}
