% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPerformance.R
\name{userPerformance}
\alias{userPerformance}
\title{userPerformance}
\usage{
userPerformance(user_id, type = "classic")
}
\arguments{
\item{user_id}{FPL user ID (retrieve from the corresponding user's fantasy.premierleague.com team page URL).}

\item{type}{"classic", "h2h" (head-to-head) or "cup". Defaults to "classic".}
}
\description{
Gives a tibble containing performance info for a given user in the current FPL season.
}
\keyword{user}
