% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPicks.R
\name{userPicks}
\alias{userPicks}
\title{Detailed data for a user's player picks for a gameweek in the current FPL season}
\usage{
userPicks(user_id, gameweek)
}
\arguments{
\item{user_id}{FPL user ID (retrieve from the corresponding user's \url{https://fantasy.premierleague.com} team page URL).}

\item{gameweek}{FPL competition gameweek (1-38).}
}
\description{
Returns a tibble containing player data for a given user's selection in a gameweek during the current FPL season.
}
\examples{
userPicks(user_id = 123, gameweek = 10)
}
